over to test

+++ nixt/booting.py
@@ -25,7 +25,6 @@ class Boot:
 
     inits = []
     md5s = {}
-    txt = " ".join(sys.argv[1:])
 
     @classmethod
     def boot(cls, name="", read=False, doall=False):
@@ -33,28 +32,18 @@ class Boot:
         Main.name = name or Main.name or Utils.pkgname(Boot)
         Main.wdr = os.path.expanduser(f"~/.{Main.name}")
         cls.core()
-        if Main.boot or read:
-            Disk.read(Main, "main", "config")
-        else:
-            Methods.parse(Main, cls.txt)
+        if Main.boot or read: Disk.read(Main, "main", "config")
+        else: Methods.parse(Main, " ".join(sys.argv[1:]))
         Workdir.skel()
         Log.size(len(Main.name))
         Log.level(Main.level or "info")
-        if Main.user:
-            Mods.add('mods', 'mods')
-        if Main.wdr:
-            Mods.add("modules", os.path.join(Main.wdr, "mods"))
-        if Main.all or doall:
-            Main.mods = Mods.list(Main.ignore)
 
     @classmethod
-    def check(cls, opts):
-        for word in cls.txt.split():
-            if not word.startswith("-"):
-                continue
+    def check(cls, opts, txt):
+        for word in txt.split():
+            if not word.startswith("-"): continue
             for char in opts:
-                if char in word:
-                    return True
+                if char in word: return True
         return False
 
     @classmethod
@@ -62,8 +51,7 @@ class Boot:
         "calculate core md5."
         path = os.path.dirname(__spec__.loader.path)
         for fnm in os.listdir(path):
-            if not fnm.endswith(".py"):
-                continue
+            if not fnm.endswith(".py"): continue
             name = fnm[:-3]
             mpath = os.path.join(path, fnm)
             cls.md5s[name] = Utils.md5sum(mpath)
@@ -72,12 +60,10 @@ class Boot:
     def daemon(cls, verbose=False, nochdir=False):
         "run in the background."
         pid = os.fork()
-        if pid != 0:
-            os._exit(0)
+        if pid != 0: os._exit(0)
         os.setsid()
         pid2 = os.fork()
-        if pid2 != 0:
-            os._exit(0)
+        if pid2 != 0: os._exit(0)
         if not verbose:
             with open('/dev/null', 'r', encoding="utf-8") as sis:
                 os.dup2(sis.fileno(), sys.stdin.fileno())
@@ -86,8 +72,7 @@ class Boot:
             with open('/dev/null', 'a+', encoding="utf-8") as ses:
                 os.dup2(ses.fileno(), sys.stderr.fileno())
         os.umask(0)
-        if not nochdir:
-            os.chdir("/")
+        if not nochdir: os.chdir("/")
         os.nice(10)
 
     @classmethod
@@ -115,8 +100,7 @@ class Boot:
     def pidfile(cls, name):
         "write pidfile."
         filename = os.path.join(Main.wdr, f"{name}.pid")
-        if os.path.exists(filename):
-            os.unlink(filename)
+        if os.path.exists(filename): os.unlink(filename)
         path2 = pathlib.Path(filename)
         path2.parent.mkdir(parents=True, exist_ok=True)
         with open(filename, "w", encoding="utf-8") as fds:
@@ -133,13 +117,11 @@ class Boot:
 
     @classmethod
     def scan(cls):
-        if Main.read:
-            cls.scanner()
+        if Main.read: cls.scanner()
         else:
             Commands.table()
             Mods.sums()
-        if not Commands.names:
-            cls.scanner()
+        if not Commands.names: cls.scanner()
 
     @classmethod
     def scanner(cls):
@@ -147,8 +129,7 @@ class Boot:
         res = []
         for name, mod in Mods.iter(Main.ignore):
             Commands.scan(mod)
-            if "configure" in dir(mod):
-                mod.configure()
+            if "configure" in dir(mod): mod.configure()
             res.append((name, mod))
         return res
